video_miner_async/pyproject.toml
[project]
name = "video-miner-v3"
version = "0.1.0"
description = "High-performance video mining pipeline for generating large-scale computer vision datasets"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Video Miner Team"}
]
keywords = ["computer-vision", "video-mining", "dataset-generation", "object-detection"]

dependencies = [
    # Core
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.40.0",
    "accelerate>=0.20.0",
    # Video/Image processing
    "opencv-python-headless>=4.8.0",
    "yt-dlp>=2024.1.0",
    "Pillow>=10.0.0",
    # ML utilities
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "faiss-cpu>=1.7.0", # Scalable similarity search
    # CLI and config
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
    # Progress and logging
    "tqdm>=4.65.0",
    "rich>=13.0.0",
    # Utilities
    "requests>=2.31.0",
    "omegaconf>=2.3.0",
    "av>=16.0.1",
    "pygount>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
]

[project.scripts]
video-miner = "video_miner_async.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["video_miner_async"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
